Bug 37362: Do not show the lists button if there are no public lists and opacuserlogin is off
This patch updates the OPAC header so that if opacuserlogin is off and
there are no public shelves, the Lists button does not appear. If
opacuserlogin is turned off the only kind of lists the user can interact
with are public ones, and without public lists there's nothing to do.

To test, apply the patch and test the arrangement of the OPAC header in
these cases:

- opacuserlogin is on, user is logged in, and there are no public lists:
  - The Lists button appears and contains only "Your lists" menu items.
- opacuserlogin is on, user is logged in, and there are public lists:
  - The Lists button appears and the menu contains controls for both
    public and private lists.
- opacuserlogin is off and there are public lists:
  - The Lists button appears and shows only controls for public lists.
- opacuserlogin is off and there are no public lists:
  - The Lists button does not appear.

Sponsored-by: Athens County Public Libraries
Signed-off-by: Caroline Cyr La Rose <caroline.cyr-la-rose@inlibro.com>
Signed-off-by: Laura_Escamilla <laura.escamilla@bywatersolutions.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>