Bug 37273: (follow-up) Remove parentheses and # from column
The new ID column shows numbers like "(#2)". I think the parentheses and
hash symbol are unnecessary and inconsistent with the way we do it in
other places and should be removed.

To test, apply the patch and rebuild the staff interface CSS: Either run
'yarn build' from the command line or if you're using a sandbox, click
the "Actions" menu for your sandbox and choose "Build CSS and JS."

Follow the previous test plan and confirm that the ID column looks
correct.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>